JavaDocs and refactoring classes
