+++ pom.xml
@@ -116,7 +116,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>com.example.demo/com.example.demo.HelloApplication</mainClass>
+                            <mainClass>com.example.demo/ES_Grupo9.HelloApplication</mainClass>
                         </configuration>
                     </execution>
                 </executions>
