+++ src/main/java/ES_Grupo9/GitManager.java
@@ -1,4 +1,4 @@
-package teste;
+package ES_Grupo9;
 
 public class GitManager{
