+++ src/main/java/ES_Grupo9/TrelloManager.java
@@ -1,4 +1,4 @@
-package teste;
+package ES_Grupo9;
 
 import org.trello4j.Trello;
 import org.trello4j.TrelloImpl;
