--- src/main/resources/com/example/demo/style.css
@@ -1,15 +1,0 @@
-.button
-{
--fx-background-color : #AED6F1;
--fx-text-fill : #827f7f;
-}
-.button:hover{
--fx-background-color : D4EAF9;
-}
-
-/*
-.button:pressed
-{
--fx-background-color : #AED6F1;
-}
-+/
